test: organize Arius.Core test doubles
