+++ src/Arius.Core.Tests/Fakes/Scenarios/ChunkStorage/BlockingDeleteBlobContainerService.cs
@@ -1,6 +1,6 @@
 using Arius.Core.Shared.Storage;
 
-namespace Arius.Core.Tests.Shared.ChunkStorage;
+namespace Arius.Core.Tests.Fakes.Scenarios.ChunkStorage;
 
 internal sealed class BlockingDeleteBlobContainerService : IBlobContainerService
 {
