+++ src/Arius.Core.Tests/Shared/ChunkStorage/ChunkStorageServiceReadTests.cs
@@ -2,6 +2,7 @@ using Arius.Core.Shared.ChunkStorage;
 using Arius.Core.Shared.Encryption;
 using Arius.Core.Shared.Storage;
 using Arius.Core.Tests.Fakes;
+using Arius.Core.Tests.Fakes.Scenarios.ChunkStorage;
 using Shouldly;
 
 namespace Arius.Core.Tests.Shared.ChunkStorage;
