+++ src/Arius.Core.Tests/Shared/FileTree/FileTreeServiceTests.cs
@@ -5,6 +5,7 @@ using Arius.Core.Shared.FileTree;
 using Arius.Core.Shared.Snapshot;
 using Arius.Core.Shared.Storage;
 using Arius.Core.Tests.Fakes;
+using Arius.Core.Tests.Fakes.Scenarios.FileTree;
 using Shouldly;
 
 namespace Arius.Core.Tests.Shared.FileTree;
